requirements.txt
streamlit
pandas
plotly
supabase
python-dotenv
altair
apify-client
google-generativeai